requirements.txt
fastapi
uvicorn
openai
python-dotenv
bcrypt
pyjwt
aiosqlite
python-multipart
httpx
